Reparent SummarizationPipeline and TranslationPipeline onto TextGenerationPipeline
