+++ src/transformers/pipelines/__init__.py
@@ -72,9 +72,9 @@ from .mask_generation import MaskGenerationPipeline
 from .object_detection import ObjectDetectionPipeline
 from .question_answering import QuestionAnsweringArgumentHandler, QuestionAnsweringPipeline
 from .table_question_answering import TableQuestionAnsweringArgumentHandler, TableQuestionAnsweringPipeline
-from .text2text_generation import SummarizationPipeline, Text2TextGenerationPipeline, TranslationPipeline
+from .text2text_generation import Text2TextGenerationPipeline
 from .text_classification import TextClassificationPipeline
-from .text_generation import TextGenerationPipeline
+from .text_generation import TextGenerationPipeline, SummarizationPipeline, TranslationPipeline
 from .text_to_audio import TextToAudioPipeline
 from .token_classification import (
     AggregationStrategy,
